+++ build_runner/lib/src/io/build_output_reader.dart
@@ -58,7 +58,7 @@ class BuildOutputReader {
   String pathFor(AssetId id) {
     return readerWriter.assetPathProvider.pathFor(
       id,
-      hide: buildState.isHidden(id),
+      inArtifactTree: buildState.isInArtifactTree(id),
     );
   }
 
@@ -95,7 +95,10 @@ class BuildOutputReader {
     }
 
     if (buildState.isSource(id) &&
-        await readerWriter.canRead(id, hidden: buildState.isHidden(id))) {
+        await readerWriter.canRead(
+          id,
+          inArtifactTree: buildState.isInArtifactTree(id),
+        )) {
       return null;
     }
     return UnreadableReason.notFound;
@@ -116,7 +119,7 @@ class BuildOutputReader {
 
     final bytes = await readerWriter.readAsBytes(
       id,
-      hidden: buildState.isHidden(id),
+      inArtifactTree: buildState.isInArtifactTree(id),
     );
     _recordSourceConsumedOutsideBuild(id);
     return BuildOutputReadResult.available(id, AssetContent.bytes(bytes));