+++ build_runner/lib/src/io/reader_writer.dart
@@ -24,17 +24,17 @@ import 'filesystem.dart';
 /// [AssetReader] and [AssetWriter] are the builder-facing file operations APIs,
 /// and are implemented here so that `TestReaderWriter` can offer them.
 ///
-/// Various methods accept `hidden`, which causes assets to be resolved under
-/// `.dart_tool/build/generated` instead of in the source tree.
+/// Various methods accept `inArtifactTree`, which causes assets to be resolved
+/// under `.dart_tool/build/generated` instead of at the package path.
 class ReaderWriter implements AssetReader, AssetWriter {
   final AssetFinder assetFinder;
   final AssetPathProvider assetPathProvider;
   final Filesystem filesystem;
 
-  /// Whether to force `hidden` to false.
+  /// Whether to force `inArtifactTree` to false.
   ///
   /// Used only in tests.
-  final bool forceVisibleForTesting;
+  final bool forceToPackagePathsForTesting;
 
   /// A [ReaderWriter] suitable for real builds.
   ///
@@ -42,48 +42,48 @@ class ReaderWriter implements AssetReader, AssetWriter {
   /// `dart-io` filesystem is used with no cache.
   factory ReaderWriter(
     BuildPackages buildPackages, {
-    bool forceVisibleForTesting = false,
+    bool forceToPackagePathsForTesting = false,
   }) => ReaderWriter.using(
     assetFinder: BuildPackagesAssetFinder(buildPackages),
     assetPathProvider: buildPackages,
     filesystem: IoFilesystem(),
-    forceVisibleForTesting: forceVisibleForTesting,
+    forceToPackagePathsForTesting: forceToPackagePathsForTesting,
   );
 
   ReaderWriter.using({
     required this.assetFinder,
     required this.assetPathProvider,
     required this.filesystem,
-    this.forceVisibleForTesting = false,
+    this.forceToPackagePathsForTesting = false,
   });
 
   String _pathFor(
     AssetId id, {
-    bool hidden = false,
+    bool inArtifactTree = false,
     bool checkWriteAllowed = false,
   }) {
     return assetPathProvider.pathFor(
       id,
-      hide: hidden && !forceVisibleForTesting,
+      inArtifactTree: inArtifactTree && !forceToPackagePathsForTesting,
       checkWriteAllowed: checkWriteAllowed,
     );
   }
 
   @override
-  Future<bool> canRead(AssetId id, {bool hidden = false}) {
+  Future<bool> canRead(AssetId id, {bool inArtifactTree = false}) {
     return Future.value(
       TimedActivity.read.run(() {
-        final path = _pathFor(id, hidden: hidden);
+        final path = _pathFor(id, inArtifactTree: inArtifactTree);
         return filesystem.existsSync(path);
       }),
     );
   }
 
   @override
-  Future<List<int>> readAsBytes(AssetId id, {bool hidden = false}) {
+  Future<List<int>> readAsBytes(AssetId id, {bool inArtifactTree = false}) {
     return Future.value(
       TimedActivity.read.run(() {
-        final path = _pathFor(id, hidden: hidden);
+        final path = _pathFor(id, inArtifactTree: inArtifactTree);
         if (!filesystem.existsSync(path)) {
           throw AssetNotFoundException(id, path: path);
         }
@@ -96,11 +96,11 @@ class ReaderWriter implements AssetReader, AssetWriter {
   Future<String> readAsString(
     AssetId id, {
     Encoding encoding = utf8,
-    bool hidden = false,
+    bool inArtifactTree = false,
   }) {
     return Future.value(
       TimedActivity.read.run(() {
-        final path = _pathFor(id, hidden: hidden);
+        final path = _pathFor(id, inArtifactTree: inArtifactTree);
         if (!filesystem.existsSync(path)) {
           throw AssetNotFoundException(id, path: path);
         }
@@ -115,10 +115,14 @@ class ReaderWriter implements AssetReader, AssetWriter {
   Future<void> writeAsBytes(
     AssetId id,
     List<int> bytes, {
-    bool hidden = false,
+    bool inArtifactTree = false,
   }) {
     TimedActivity.write.run(() {
-      final path = _pathFor(id, hidden: hidden, checkWriteAllowed: true);
+      final path = _pathFor(
+        id,
+        inArtifactTree: inArtifactTree,
+        checkWriteAllowed: true,
+      );
       filesystem.writeAsBytesSync(path, bytes);
     });
     return Future.value();
@@ -129,36 +133,48 @@ class ReaderWriter implements AssetReader, AssetWriter {
     AssetId id,
     String contents, {
     Encoding encoding = utf8,
-    bool hidden = false,
+    bool inArtifactTree = false,
   }) {
     TimedActivity.write.run(() {
-      final path = _pathFor(id, hidden: hidden, checkWriteAllowed: true);
+      final path = _pathFor(
+        id,
+        inArtifactTree: inArtifactTree,
+        checkWriteAllowed: true,
+      );
       filesystem.writeAsStringSync(path, contents, encoding: encoding);
     });
     return Future.value();
   }
 
   @override
-  Future<Digest> digest(AssetId id, {bool hidden = false}) async {
+  Future<Digest> digest(AssetId id, {bool inArtifactTree = false}) async {
     final digestSink = AccumulatorSink<Digest>();
     md5.startChunkedConversion(digestSink)
-      ..add(await readAsBytes(id, hidden: hidden))
+      ..add(await readAsBytes(id, inArtifactTree: inArtifactTree))
       ..add(id.toString().codeUnits)
       ..close();
     return digestSink.events.first;
   }
 
-  Future<void> delete(AssetId id, {bool hidden = false}) {
+  Future<void> delete(AssetId id, {bool inArtifactTree = false}) {
     TimedActivity.write.run(() {
-      final path = _pathFor(id, hidden: hidden, checkWriteAllowed: true);
+      final path = _pathFor(
+        id,
+        inArtifactTree: inArtifactTree,
+        checkWriteAllowed: true,
+      );
       filesystem.deleteSync(path);
     });
     return Future.value();
   }
 
-  Future<void> deleteDirectory(AssetId id, {bool hidden = false}) {
+  Future<void> deleteDirectory(AssetId id, {bool inArtifactTree = false}) {
     TimedActivity.write.run(() {
-      final path = _pathFor(id, hidden: hidden, checkWriteAllowed: true);
+      final path = _pathFor(
+        id,
+        inArtifactTree: inArtifactTree,
+        checkWriteAllowed: true,
+      );
       filesystem.deleteDirectorySync(path);
     });
     return Future.value();