+++ build_runner/test/io/build_output_reader_test.dart
@@ -123,7 +123,7 @@ void main() {
       buildState.addPostProcessBuildStepResult(
         step: PostProcessBuildStepId(input: deletedId, actionNumber: 0),
         result: PostProcessBuildStepResult(
-          hidden: true,
+          inArtifactTree: true,
           deletedPrimaryInput: true,
         ),
       );
@@ -205,7 +205,7 @@ void main() {
       );
       final stepResult = BuildStepResult((b) {
         b.result = false;
-        b.isHidden = false;
+        b.inArtifactTree = false;
       });
       buildState.addBuildStepResult(step: buildStepId, result: stepResult);
       reader = BuildOutputReader(
@@ -267,7 +267,7 @@ void main() {
       buildState.addPostProcessBuildStepResult(
         step: postProcessId,
         result: PostProcessBuildStepResult(
-          hidden: false,
+          inArtifactTree: false,
           deletedPrimaryInput: false,
           outputs: [postOutput],
         ),
@@ -358,7 +358,7 @@ void main() {
         buildState.addPostProcessBuildStepResult(
           step: postProcessId,
           result: PostProcessBuildStepResult(
-            hidden: false,
+            inArtifactTree: false,
             deletedPrimaryInput: false,
             outputs: [generatedId],
           ),
@@ -444,7 +444,7 @@ void main() {
         );
         final stepResult = BuildStepResult((b) {
           b.result = false;
-          b.isHidden = false;
+          b.inArtifactTree = false;
         });
         buildState.addBuildStepResult(step: buildStepId, result: stepResult);
         reader = BuildOutputReader(
@@ -504,7 +504,7 @@ void main() {
       buildState.addSourceForTest(readableId);
       final stepResult = BuildStepResult((b) {
         b.result = false;
-        b.isHidden = false;
+        b.inArtifactTree = false;
       });
       buildState.addBuildStepResult(step: failedStepId, result: stepResult);
       reader = BuildOutputReader(